backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
langchain==0.1.9
chromadb==0.4.22
sentence-transformers==2.5.1
together==0.2.11
pypdf==4.0.2
pydantic==2.6.1
langchain-community
pypdf
python-dotenv
langchain_together